refactor: enhance depth estimation client with batch processing and caching
- Implemented batch processing for food portion estimation, allowing multiple foods to be processed in a single API call, significantly improving efficiency.
- Added image hash caching to avoid reprocessing the same images, enhancing performance and reducing redundant API calls.
- Updated the depth estimation client to support both individual and batch estimation methods, streamlining the overall workflow.

+++ kai/agents/florence_bbox.py
@@ -61,7 +61,7 @@ class FlorenceBBoxDetector:
             dtype = torch.float16 if self.device == "cuda" else torch.float32
             self.model = AutoModelForCausalLM.from_pretrained(
                 self.model_name,
-                torch_dtype=dtype,  # Use torch_dtype (transformers expects this)
+                dtype=dtype,  # Use dtype (torch_dtype is deprecated)
                 trust_remote_code=True,
                 attn_implementation="eager"  # Fix for Python 3.13: disable SDPA
             )
